Changed the type data to double

+++ src/Calculate.java
@@ -1,5 +1,5 @@
 public class Calculate {
-    public static int  Calculating(String choice, int a, int b){
+    public static double  Calculating(String choice, double a, double b){
         if(choice == "Add"){
             return Add(a,b);
         }else if(choice == "Substract"){
@@ -13,16 +13,16 @@ public class Calculate {
         return 0;
     }
 
-    public static int Add(int a, int b){
+    public static  double Add(double a, double b){
         return a+b;
     }
-    public static int Substract(int a, int b){
+    public static double Substract(double a, double b){
         return a-b;
     }
-    public static int Multiply(int a, int b){
+    public static double Multiply(double a, double b){
         return a*b;
     }
-    public static int Divide(int a, int b){
+    public static double Divide(double a, double b){
         return a/b;
     }
 }
